Refactor tool call middleware tests by consolidating positive path scenarios
- Moved the positive path test for `createToolMiddleware` from a separate file to the main test suite, enhancing organization and maintainability.
- This change improves the clarity of the testing framework by consolidating related tests and ensuring better coverage for the tool call middleware functionality.
